+++ crml/support/src/lib.rs
@@ -274,5 +274,5 @@ pub trait StakingInfo {
 	/// Gets the active balance of a controller accounts staked amount
 	fn active_balance(controller: Self::AccountId) -> Self::Balance;
 	/// Counts the current amount of nominators
-	fn count_nominators() -> u64;
+	fn count_nominators() -> u32;
 }
